Incremental Index
The incremental engine in MPS-extensions doesn't have an
efficient way to find a transformation output node.
This would be required for the shadow models based editor.

The new data structures allow to update the HashMap of the
index incrementally while the function that defines the content
of the index still uses dynamic dependency analysis.