SLRU cache eviction algorithm
The graph size limit should be proportional to
the working set and not to the model size.
The benchmark RecursiveLargeSum1_000 checks this.
It uses a model with 1_000_000 entries, but only 10
entries are modified.
The issue was that computing a subgraph created more
nodes in the dependency graph than the size limit
causing all other nodes being evicted.
The SLRU cache has a separate segment for entries
that were used only once. These short living nodes
then only replace other short living ones instead of
the useful part of the cache entries.